Add Bus and Stop data classes; update User for ownerUid and assignedBusId
